split stories into 2 file
